input/agent.rs: Split out objective-related code into submodule
